Modified redux implementation, mainly refactoring code
