+++ src/features/driverform/store/state.js
@@ -1,4 +1,8 @@
-const initialState = {};
+const initialState = {
+  AllDrivers: [],
+  LastDriver: {},
+  error: false
+};
 
 export {
   initialState
