--- src/stores/consts.js
@@ -1,7 +1,0 @@
-const SAVE_DRIVER = '[DRIVER_FORM] SAVE_DRIVER';
-const SAVE_ALL_DRIVERS = '[DRIVER_FORM] SAVE_ALL_DRIVERS';
-
-export {
-  SAVE_DRIVER,
-  SAVE_ALL_DRIVERS
-}
