+++ src/stores/index.js
@@ -1,9 +1,12 @@
-import { createStore, applyMiddleware } from 'redux';
+import {
+  createStore,
+  applyMiddleware
+} from 'redux';
 import thunk from 'redux-thunk';
-import { Driver } from './reducers';
+import { rootReducer } from './reducers';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
-const store = createStore(Driver, composeWithDevTools(applyMiddleware(thunk)));
+const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
 
 export {
   store
