+++ src/stores/reducers.js
@@ -1,28 +1,11 @@
-import { initialState } from './state';
-import {
-  SAVE_DRIVER,
-  SAVE_ALL_DRIVERS
-} from './consts';
+import { combineReducers } from 'redux';
+import { driversReducers } from '../features/driverform/store/reducers';
 
-const Driver = (state = initialState, action) => {
-  switch (action.type) {
-    case SAVE_DRIVER: {
-      const { payload: { response } } = action;
-      return Object.assign({}, state, {
-        LastDriver: response
-      });
-    }
-    case SAVE_ALL_DRIVERS: {
-      const { payload: { response } } = action;
-      return Object.assign({}, state, {
-        AllDrivers: [...state.AllDrivers, response]
-      });
-    }
-    default:
-      return state
-  }
-}
+const rootReducer = combineReducers({
+  driversReducers
+});
 
 export {
-  Driver
+  rootReducer
 }
+
