--- src/stores/state.js
@@ -1,9 +1,0 @@
-const initialState = {
-  AllDrivers: [],
-  LastDriver: {},
-  error: false
-};
-
-export {
-  initialState
-}
